+++ NEWS.md
@@ -1,3 +1,27 @@
+SqlRender 1.6.2
+===============
+
+Changes:
+
+1. Added rules for Oracle for conditonal indices.
+
+Bugfixes:
+
+1. Fixing erroneous variable name translation for BigQuery.
+
+
+SqlRender 1.6.1
+===============
+
+Changes:
+
+1. Added rules for Impala for INTEGER NOT NULL and DOUBLE PRECISION.
+
+Bugfixes:
+
+1. Fixed isNumeric check for Netezza
+
+
 SqlRender 1.6.0
 ===============